requirements.txt
discord.py
andyontop